Moving state out of main mission planner adapter. New settings.

+++ ExtLibs/AltitudeAngelWings/ApiClient/Client/AltitudeAngelClient.cs
@@ -229,7 +229,7 @@ namespace AltitudeAngelWings.ApiClient.Client
 
         private static CreateFlightPartRequest CreateFlightPartRequest(
             int id,
-            int highestAltitude,
+            double highestAltitude,
             Instant start,
             Instant end,
             Feature geography)
