+++ ExtLibs/AltitudeAngelWings/Models/FlightPlanWaypoint.cs
@@ -4,6 +4,6 @@ namespace AltitudeAngelWings.Models
     {
         public double Longitude { get; set; }
         public double Latitude { get; set; }
-        public int Altitude { get; set; }
+        public double Altitude { get; set; }
     }
 }
